Fix compiler warnings in MainComponent, LookAndFeel, and LevelMeter
- Fix destructor override and extra semicolons in MelissaLookAndFeel.h
- Fix implicit int-to-float and float-conversion warnings throughout
- Replace deprecated M_PI with MathConstants<float>::pi
- Replace deprecated Font(float) with Font(FontOptions().withHeight())
- Replace deprecated getStringWidth with getStringWidthFloat
- Remove unused virtual keywords and comment out unused parameters
- Fix shadow variable warnings in resized() by renaming inner 'y' to 'iy'
- Fix int/size_t and sign conversion warnings in audio callbacks
- Fix int-to-float conversions in LevelMeter calculations

+++ Source/UI/MelissaLevelMeter.h
@@ -69,8 +69,8 @@ public:
             rmsL += left[i]  * left[i];
             rmsR += right[i] * right[i];
         }
-        rmsL = std::sqrt(rmsL / numSamples);
-        rmsR = std::sqrt(rmsR / numSamples);
+        rmsL = std::sqrt(rmsL / static_cast<float>(numSamples));
+        rmsR = std::sqrt(rmsR / static_cast<float>(numSamples));
 
         auto blend = [](float cur, float next) {
             return next > cur ? next : cur * 0.85f + next * 0.15f;
@@ -166,7 +166,7 @@ private:
 
     void setValueFromX(int x)
     {
-        const float barW = getWidth() - kBarMargin * 2.f;
+        const float barW = static_cast<float>(getWidth()) - kBarMargin * 2.f;
         const float v = juce::jlimit(0.01f, 1.f, (static_cast<float>(x) - kBarMargin) / barW);
         setValue(v, true);
     }
